Removing progress-bar.js
